# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server API for Hamster Money Service Application",
        "title": "Hamster API",
        "contact": {
            "name": "Hamster API Support",
            "url": "https://t.me/CodeMaster482",
            "email": "grigorikovalenko@gmail.com"
        },
        "version": "1.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/user/{userID}/account/feed",
    "paths": {
...
        "/api/user/": {
            "get": {
                "description": "Get user with chosen ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get User",
                "responses": {
                    "200": {
                        "description": "Show user",
                        "schema": {
                            "$ref": "#/definitions/http.Response-transfer_models_UserTransfer"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/user/accounts/all": {
            "get": {
                "description": "Get User accounts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get User Accounts",
                "responses": {
                    "200": {
                        "description": "Show actual accounts",
                        "schema": {
                            "$ref": "#/definitions/http.Response-transfer_models_Account"
                        }
                    },
                    "204": {
                        "description": "Show actual accounts",
                        "schema": {
                            "$ref": "#/definitions/http.Response-string"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/user/actualBudget": {
            "get": {
                "description": "Get User actual budget",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get Actual Budget",
                "responses": {
                    "200": {
                        "description": "Show actual budget",
                        "schema": {
                            "$ref": "#/definitions/http.Response-transfer_models_BudgetActualResponse"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/user/balance": {
            "get": {
                "description": "Get User balance",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get Balance",
                "responses": {
                    "200": {
                        "description": "Show balance",
                        "schema": {
                            "$ref": "#/definitions/http.Response-transfer_models_BalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/user/feed": {
            "get": {
                "description": "Get Feed user info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get Feed",
                "responses": {
                    "200": {
                        "description": "Show actual accounts",
                        "schema": {
                            "$ref": "#/definitions/http.Response-transfer_models_UserFeed"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/user/plannedBudget": {
            "get": {
                "description": "Get User planned budget",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get Planned Budget",
                "responses": {
                    "200": {
                        "description": "Show planned budget",
                        "schema": {
                            "$ref": "#/definitions/http.Response-transfer_models_BudgetPlannedResponse"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/user/update": {
            "put": {
                "description": "Update user info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "PUT Update",
                "parameters": [
                    {
                        "description": "user info update",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transfer_models.UserUdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update user info",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_NilBody"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/user/updatePhoto": {
            "put": {
                "description": "Update user photo",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "PUT Update Photo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "New photo to upload",
                        "name": "upload",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Path to old photo",
                        "name": "path",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Photo updated successfully",
                        "schema": {
                            "$ref": "#/definitions/http.Response-transfer_models_PhotoUpdate"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "http.NilBody": {
            "type": "object"
        },
...
        "http.Response-http_NilBody": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/http.NilBody"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "http.Response-string": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "http.Response-transfer_models_Account": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/transfer_models.Account"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "http.Response-transfer_models_BalanceResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/transfer_models.BalanceResponse"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "http.Response-transfer_models_BudgetActualResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/transfer_models.BudgetActualResponse"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "http.Response-transfer_models_BudgetPlannedResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/transfer_models.BudgetPlannedResponse"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "http.Response-transfer_models_PhotoUpdate": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/transfer_models.PhotoUpdate"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "http.Response-transfer_models_UserFeed": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/transfer_models.UserFeed"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "http.Response-transfer_models_UserTransfer": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/transfer_models.UserTransfer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "http.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "models.Accounts": {
            "type": "object",
            "properties": {
                "accumulation": {
                    "type": "boolean"
                },
                "balance": {
                    "type": "number"
                },
                "balance_enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "mean_payment": {
                    "type": "string"
                }
            }
        },
...
        "transfer_models.Account": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Accounts"
                    }
                }
            }
        },
        "transfer_models.BalanceResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                }
            }
        },
        "transfer_models.BudgetActualResponse": {
            "type": "object",
            "properties": {
                "actual_budget": {
                    "type": "number"
                }
            }
        },
        "transfer_models.BudgetPlannedResponse": {
            "type": "object",
            "properties": {
                "planned_budget": {
                    "type": "number"
                }
            }
        },
        "transfer_models.PhotoUpdate": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
        "transfer_models.UserFeed": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Accounts"
                    }
                },
                "actual_budget": {
                    "type": "number"
                },
                "balance": {
                    "type": "number"
                },
                "planned_budget": {
                    "type": "number"
                }
            }
        },
        "transfer_models.UserTransfer": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "planned_budget": {
                    "type": "number"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "transfer_models.UserUdate": {
            "type": "object",
            "properties": {
                "planned_budget": {
                    "type": "number"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}